+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+* Don't delete the `srcset` attribute from `<img>`.
 * Embed style sheets in single-file mode using data URIs rather than `<style>`.
 ### Fixed
 * Handle `srcset` entries without a width or pixel density descriptor.